+++ eslint.config.mjs
@@ -14,6 +14,7 @@ export default moomfe(
     rules: {
       'no-console': 'off',
       'style/jsx-one-expression-per-line': 'off',
+      'vue/no-template-shadow': 'off',
     },
   },
 );